refactor: restructure image generation plugin by separating logic into dedicated action and utility files
