+++ packages/plugin-image-generation/src/utils/index.ts
@@ -1,0 +1,51 @@
+import fs from "fs";
+import path from "path";
+
+export function saveBase64Image(base64Data: string, filename: string): string {
+    // Create generatedImages directory if it doesn't exist
+    const imageDir = path.join(process.cwd(), "generatedImages");
+    if (!fs.existsSync(imageDir)) {
+        fs.mkdirSync(imageDir, { recursive: true });
+    }
+
+    // Remove the data:image/png;base64 prefix if it exists
+    const base64Image = base64Data.replace(/^data:image\/\w+;base64,/, "");
+
+    // Create a buffer from the base64 string
+    const imageBuffer = Buffer.from(base64Image, "base64");
+
+    // Create full file path
+    const filepath = path.join(imageDir, `${filename}.png`);
+
+    // Save the file
+    fs.writeFileSync(filepath, imageBuffer);
+
+    return filepath;
+}
+
+export async function saveHeuristImage(
+    imageUrl: string,
+    filename: string
+): Promise<string> {
+    const imageDir = path.join(process.cwd(), "generatedImages");
+    if (!fs.existsSync(imageDir)) {
+        fs.mkdirSync(imageDir, { recursive: true });
+    }
+
+    // Fetch image from URL
+    const response = await fetch(imageUrl);
+    if (!response.ok) {
+        throw new Error(`Failed to fetch image: ${response.statusText}`);
+    }
+
+    const arrayBuffer = await response.arrayBuffer();
+    const imageBuffer = Buffer.from(arrayBuffer);
+
+    // Create full file path
+    const filepath = path.join(imageDir, `${filename}.png`);
+
+    // Save the file
+    fs.writeFileSync(filepath, imageBuffer);
+
+    return filepath;
+}
